+++ src/aks-preview/HISTORY.rst
@@ -17,12 +17,23 @@ Pending
 * Vendor new SDK and bump API version to 2023-05-02-preview.
 * Add support for new snapshot command `az aks nodepool snapshot update`
 
+0.5.145
++++++++
+* Add support for option --aks-custom-headers to some aks commands
+  * aks get-credentials
+  * aks nodepool scale
+  * aks nodepool update
+  * aks enable-addons
+  * aks show
+  * aks scale
+
 0.5.144
 +++++++
 * Fix setup network profile with network observability due to incorrect property
 
 0.5.143
 +++++++
+* Vendor new SDK and bump API version to 2023-05-02-preview.
 * Add `--enable-network-observability` flag to `az aks create` and `az aks update`.
 
 0.5.142